k8s_mlops/ml-training/pyproject.toml
[project]
name = "wine-quality-training-pipeline"
version = "0.1.0"
description = "Enterprise-style ML training pipeline for wine quality classification on Kubernetes"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "scikit-learn>=1.5",
    "optuna>=3.6",
    "numpy>=1.26",
    "pandas>=2.2",
    "joblib>=1.4",
    "pyyaml>=6.0",
]

[project.scripts]
run-training-pipeline = "wine_quality_training.pipeline.run_training_pipeline:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/wine_quality_training"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]
